Moved parseAxisParams to util.ts
parseAxisParams which caused a dependency loop is moved to util.ts
All the imports and references of the same have been updated
