+++ src/ops/axis_util.ts
@@ -65,29 +65,6 @@ export function expandShapeToKeepDim(
   return combineLocations(shape, reduceSubShape, axes);
 }
 
-export function parseAxisParam(
-    axis: number|number[], shape: number[]): number[] {
-  const rank = shape.length;
-
-  // Normalize input
-  axis = axis == null ? shape.map((s, i) => i) : [].concat(axis);
-
-  // Check for valid range
-  util.assert(
-      axis.every(ax => ax >= -rank && ax < rank),
-      `All values in axis param must be in range [-${rank}, ${rank}) but ` +
-          `got axis ${axis}`);
-
-  // Check for only integers
-  util.assert(
-      axis.every(ax => util.isInt(ax)),
-      `All values in axis param must be integers but ` +
-          `got axis ${axis}`);
-
-  // Handle negative axis.
-  return axis.map(a => a < 0 ? rank + a : a);
-}
-
 export function assertAxesAreInnerMostDims(
     msg: string, axes: number[], rank: number): void {
   util.assert(
