+++ src/ops/reduction_ops.ts
@@ -57,7 +57,7 @@ function logSumExp_<T extends Tensor>(
     x: Tensor|TensorLike, axis: number|number[] = null, keepDims = false): T {
   const $x = convertToTensor(x, 'x', 'logSumExp');
 
-  const axes = axis_util.parseAxisParam(axis, $x.shape);
+  const axes = util.parseAxisParam(axis, $x.shape);
   const xMax = $x.max(axes, true /* keepDims */);
   const a = $x.sub(xMax);
   const b = a.exp();
@@ -108,7 +108,7 @@ function sum_<T extends Tensor>(
   if ($x.dtype === 'bool') {
     $x = $x.toInt();
   }
-  const axes = axis_util.parseAxisParam(axis, $x.shape);
+  const axes = util.parseAxisParam(axis, $x.shape);
 
   // Use a custom gradient to bypass 2 gradient backprops since sum is used
   // extremely often.
@@ -178,7 +178,7 @@ function prod_<T extends Tensor>(
   if ($x.dtype === 'bool') {
     $x = $x.toInt();
   }
-  const axes = axis_util.parseAxisParam(axis, $x.shape);
+  const axes = util.parseAxisParam(axis, $x.shape);
 
   const permutation = axis_util.getAxesPermutation(axes, $x.rank);
   let reductionAxes = axes;
@@ -228,7 +228,7 @@ function mean_<T extends Tensor>(
     x: Tensor|TensorLike, axis: number|number[] = null, keepDims = false): T {
   const $x = convertToTensor(x, 'x', 'mean');
 
-  const axes = axis_util.parseAxisParam(axis, $x.shape);
+  const axes = util.parseAxisParam(axis, $x.shape);
   const shapes = axis_util.computeOutAndReduceShapes($x.shape, axes);
   const reduceShape = shapes[1];
   const reduceSize = util.sizeFromShape(reduceShape);
@@ -313,7 +313,7 @@ function min_<T extends Tensor>(
   let $x = convertToTensor(x, 'x', 'min');
   const xOrig = $x;
 
-  const origAxes = axis_util.parseAxisParam(axis, $x.shape);
+  const origAxes = util.parseAxisParam(axis, $x.shape);
   let axes = origAxes;
   const permutedAxes = axis_util.getAxesPermutation(axes, $x.rank);
   if (permutedAxes != null) {
@@ -365,7 +365,7 @@ function max_<T extends Tensor>(
   let $x = convertToTensor(x, 'x', 'max');
   const xOrig = $x;
 
-  const origAxes = axis_util.parseAxisParam(axis, $x.shape);
+  const origAxes = util.parseAxisParam(axis, $x.shape);
   let axes = origAxes;
   const permutedAxes = axis_util.getAxesPermutation(axes, $x.rank);
   if (permutedAxes != null) {
@@ -414,7 +414,7 @@ function argMin_<T extends Tensor>(x: Tensor|TensorLike, axis = 0): T {
   if (axis == null) {
     axis = 0;
   }
-  let axes = axis_util.parseAxisParam(axis, $x.shape);
+  let axes = util.parseAxisParam(axis, $x.shape);
   const permutedAxes = axis_util.getAxesPermutation(axes, $x.rank);
   if (permutedAxes != null) {
     $x = $x.transpose(permutedAxes);
@@ -456,7 +456,7 @@ function argMax_<T extends Tensor>(x: Tensor|TensorLike, axis = 0): T {
   if (axis == null) {
     axis = 0;
   }
-  let axes = axis_util.parseAxisParam(axis, $x.shape);
+  let axes = util.parseAxisParam(axis, $x.shape);
   const permutedAxes = axis_util.getAxesPermutation(axes, $x.rank);
   if (permutedAxes != null) {
     $x = $x.transpose(permutedAxes);
@@ -501,7 +501,7 @@ function all_<T extends Tensor>(
     x: Tensor|TensorLike, axis: number|number[] = null, keepDims = false): T {
   let $x = convertToTensor(x, 'x', 'all', 'bool');
 
-  const origAxes = axis_util.parseAxisParam(axis, $x.shape);
+  const origAxes = util.parseAxisParam(axis, $x.shape);
   let axes = origAxes;
   const permutedAxes = axis_util.getAxesPermutation(axes, $x.rank);
   if (permutedAxes != null) {
@@ -548,7 +548,7 @@ function any_<T extends Tensor>(
     x: Tensor|TensorLike, axis: number|number[] = null, keepDims = false): T {
   let $x = convertToTensor(x, 'x', 'any', 'bool');
 
-  const origAxes = axis_util.parseAxisParam(axis, $x.shape);
+  const origAxes = util.parseAxisParam(axis, $x.shape);
   let axes = origAxes;
   const permutedAxes = axis_util.getAxesPermutation(axes, $x.rank);
   if (permutedAxes != null) {
@@ -580,7 +580,7 @@ function moments_(
     x: Tensor|TensorLike, axis: number|number[] = null,
     keepDims = false): {mean: Tensor, variance: Tensor} {
   x = convertToTensor(x, 'x', 'moments');
-  const axes = axis_util.parseAxisParam(axis, x.shape);
+  const axes = util.parseAxisParam(axis, x.shape);
   const mean = x.mean(axes, keepDims);
   let keepDimsShape = mean.shape;
   if (!keepDims) {
