+++ src/ops/segment_ops.ts
@@ -19,9 +19,9 @@ import {ENV} from '../environment';
 import {Tensor, Tensor1D} from '../tensor';
 import {convertToTensor} from '../tensor_util_env';
 import {TensorLike} from '../types';
-import {assert, isInt} from '../util';
+import {assert, isInt, parseAxisParam} from '../util';
 import {expandDims} from './array_ops';
-import {getUndoAxesPermutation, parseAxisParam} from './axis_util';
+import {getUndoAxesPermutation} from './axis_util';
 import {maximum} from './binary_ops';
 import {greaterEqual} from './compare';
 import {logicalAnd, where} from './logical_ops';
