+++ src/util.ts
@@ -16,7 +16,6 @@
  */
 
 import {DataType, DataTypeMap, FlatVector, NumericDataType, RecursiveArray, TensorLike, TypedArray} from './types';
-import {parseAxisParam} from './ops/axis_util';
 
 /** Shuffles the array using Fisher-Yates algorithm. */
 // tslint:disable-next-line:no-any
@@ -269,6 +268,29 @@ export function inferFromImplicitShape(
   const newShape = shape.slice();
   newShape[implicitIdx] = size / shapeProd;
   return newShape;
+}
+
+export function parseAxisParam(
+    axis: number|number[], shape: number[]): number[] {
+  const rank = shape.length;
+
+  // Normalize input
+  axis = axis == null ? shape.map((s, i) => i) : [].concat(axis);
+
+  // Check for valid range
+  assert(
+      axis.every(ax => ax >= -rank && ax < rank),
+      `All values in axis param must be in range [-${rank}, ${rank}) but ` +
+          `got axis ${axis}`);
+
+  // Check for only integers
+  assert(
+      axis.every(ax => isInt(ax)),
+      `All values in axis param must be integers but ` +
+          `got axis ${axis}`);
+
+  // Handle negative axis.
+  return axis.map(a => a < 0 ? rank + a : a);
 }
 
 /** Reduces the shape by removing all dimensions of shape 1. */
